requirements.txt
bokeh==2.4.0
matplotlib==3.4.3
numpy==1.19.5
opencv_contrib_python==4.5.3.56
pandas==1.3.3
Pillow==9.0.0
pytorch_pretrained_vit==0.0.7
requests==2.26.0
scikit_image==0.18.3
scikit_learn==1.0.2
scipy==1.7.1
scikit-image
submitit==1.4.1
torch==1.9.1+cu111
torchattacks==3.2.3
torchvision==0.10.1+cu111
tqdm==4.62.3
umap==0.1.1
umap_learn==0.5.2
json==2.0.9
